Get Style and Compilation of Lifestuff on Windows [#18023179]

+++ src/maidsafe/lifestuff/contacts.cc
@@ -75,7 +75,7 @@ int ContactsHandler::AddContact(const std::string &pub_name,
                                 const uint32_t &last_contact) {
   uint32_t lc = 0;
   if (last_contact == 0)
-    lc = GetDurationSinceEpoch().total_milliseconds();
+    lc = static_cast<uint32_t>(GetDurationSinceEpoch().total_milliseconds());
   else
     lc = last_contact;
   mi_contact mic(pub_name, pub_key, full_name, office_phone, birthday, gender,
@@ -253,7 +253,8 @@ int ContactsHandler::SetLastContactRank(const std::string &pub_name) {
   }
   mi_contact mic = *it;
   mic.rank_++;
-  mic.last_contact_ = GetDurationSinceEpoch().total_milliseconds();
+  mic.last_contact_ =
+      static_cast<uint32_t>(GetDurationSinceEpoch().total_milliseconds());
   cs_.replace(it, mic);
   return 0;
 }
